+++ packages/types/src/config/config.ts
@@ -33,6 +33,7 @@ import {
 	FrictionlessPenalties,
 	PENALTY_ACCESS_RULE_DEFAULT,
 	PENALTY_OLD_TIMESTAMP_DEFAULT,
+	PENALTY_UNVERIFIED_HOST_DEFAULT,
 } from "./frictionless.js";
 import {
 	DEFAULT_IMAGE_CAPTCHA_SOLUTION_TIMEOUT,
@@ -272,6 +273,7 @@ export type ProcaptchaClientConfigOutput = output<
 
 export const ProsopoConfigSchema = ProsopoBasicConfigSchema.merge(
 	object({
+		host: string(),
 		captchas: ProsopoCaptchaCountConfigSchema.optional().default({
 			solved: { count: DEFAULT_SOLVED_COUNT },
 			unsolved: { count: DEFAULT_UNSOLVED_COUNT },
@@ -279,6 +281,7 @@ export const ProsopoConfigSchema = ProsopoBasicConfigSchema.merge(
 		penalties: FrictionlessPenalties.optional().default({
 			PENALTY_OLD_TIMESTAMP: PENALTY_OLD_TIMESTAMP_DEFAULT,
 			PENALTY_ACCESS_RULE: PENALTY_ACCESS_RULE_DEFAULT,
+			PENALTY_UNVERIFIED_HOST: PENALTY_UNVERIFIED_HOST_DEFAULT,
 		}),
 		scheduledTasks: object({
 			captchaScheduler: object({