+++ packages/dotenv/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @prosopo/dotenv
 
+## 3.0.14
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [809b984]
+  - @prosopo/common@3.1.9
+  - @prosopo/config@3.1.9
+
+## 3.0.13
+### Patch Changes
+
+- 6fe8570: Release 3.2.0
+- Updated dependencies [6fe8570]
+  - @prosopo/common@3.1.8
+  - @prosopo/config@3.1.8
+
+## 3.0.12
+### Patch Changes
+
+- f304be9: Release 3.1.13
+- Updated dependencies [f304be9]
+  - @prosopo/common@3.1.7
+  - @prosopo/config@3.1.7
+
+## 3.0.11
+### Patch Changes
+
+- Updated dependencies [9eed772]
+  - @prosopo/config@3.1.6
+  - @prosopo/common@3.1.6
+
 ## 3.0.10
 ### Patch Changes